split and move structures from jormungandr scenarios tests to jormungandr lib. All abstract structures were moved to libs. if structures had some ui or thread dependency (progress bar, tokio) they were splitted.
