+++ jormungandr-lib/src/testing/mod.rs
@@ -2,6 +2,7 @@ mod archive;
 pub mod github;
 /// Module contains cross project test utils
 mod measurement;
+pub mod network_builder;
 mod web;
 
 pub use archive::decompress;
@@ -11,4 +12,5 @@ pub use measurement::{
     EnduranceBenchmarkDef, EnduranceBenchmarkFinish, EnduranceBenchmarkRun, Speed,
     SpeedBenchmarkDef, SpeedBenchmarkFinish, SpeedBenchmarkRun, Thresholds, Timestamp,
 };
+
 pub use web::download_file;
