+++ jormungandr-scenario-tests/src/test/network/real.rs
@@ -1,9 +1,8 @@
 use crate::{
     node::{LeadershipMode, PersistenceMode},
     scenario::{
-        repository::ScenarioResult, ActiveSlotCoefficient, Blockchain, ConsensusVersion,
-        ControllerBuilder, KESUpdateSpeed, Milli, Node, NumberOfSlotsPerEpoch, SlotDuration,
-        TopologyBuilder, Value, Wallet,
+        repository::ScenarioResult, ActiveSlotCoefficient, ConsensusVersion, ControllerBuilder,
+        KESUpdateSpeed, Milli, Node, NumberOfSlotsPerEpoch, SlotDuration, Value,
     },
     test::{
         utils::{self, SyncMeasurementInterval, SyncWaitParams},
@@ -11,6 +10,8 @@ use crate::{
     },
     Context,
 };
+
+use jormungandr_lib::testing::network_builder::{Blockchain, TopologyBuilder, WalletTemplate};
 
 use rand_chacha::ChaChaRng;
 
@@ -78,7 +79,8 @@ fn prepare_real_scenario(
 
     for i in 1..leader_counter {
         let initial_wallet_name = wallet_name(i);
-        let mut wallet = Wallet::new_account(initial_wallet_name.to_owned(), Value(100_000));
+        let mut wallet =
+            WalletTemplate::new_account(initial_wallet_name.to_owned(), Value(100_000).into());
         *wallet.delegate_mut() = Some(leader_name(i).to_owned());
         blockchain.add_wallet(wallet);
     }
